+++ pandas/plotting/_misc.py
@@ -68,7 +68,7 @@ def table(ax: Axes, data: DataFrame | Series, **kwargs) -> Table:
             >>> df = pd.DataFrame({"A": [1, 2], "B": [3, 4]})
             >>> fig, ax = plt.subplots()
             >>> ax.axis("off")
-            (0.0, 1.0, 0.0, 1.0)
+            (np.float64(0.0), np.float64(1.0), np.float64(0.0), np.float64(1.0))
             >>> table = pd.plotting.table(
             ...     ax, df, loc="center", cellLoc="center", colWidths=[0.2, 0.2]
             ... )
